src/requirements.txt
APScheduler
Flask==3.0.3
openpyxl
pandas
paramiko==3.4.0
pillow
pymssql
py-healthcheck
pysftp
psycopg2
prometheus-client
python-dotenv
requests==2.31.0
requests-pkcs12==1.24
SQLAlchemy
streamlit
streamlit-tree-select
streamlit-antd-components
streamlit-keycloak-zpl
https://github.com/Randers-Kommune-Digitalisering/streamlit-flow/releases/download/build-run5/streamlit_flow_component-1.7.0-py3-none-any.whl